api: move original code to .old package
