+++ src/main/java/com/revtekk/nioflex/old/NIOServer.java
@@ -1,7 +1,7 @@
-package com.revtekk.nioflex;
+package com.revtekk.nioflex.old;
 
-import com.revtekk.nioflex.policy.ThreadPolicy;
-import com.revtekk.nioflex.utils.SocketUtil;
+import com.revtekk.nioflex.old.policy.ThreadPolicy;
+import com.revtekk.nioflex.old.utils.SocketUtil;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
