+++ src/main/java/com/revtekk/nioflex/old/utils/SocketUtil.java
@@ -1,4 +1,4 @@
-package com.revtekk.nioflex.utils;
+package com.revtekk.nioflex.old.utils;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
